refactor: iOS Gallery refresh files change
